requirements.txt
# Dependencias del proyecto IA_EVO_004_01

# Para cargar variables de entorno desde el archivo .env
python-dotenv

# Librerías para los clientes de API de IA
google-generativeai
ollama

#Para verificar modelos de ollama
requests

# Para la funcionalidad RAG
pypdf
chromadb

# Langchain para capacidades avanzadas de procesamiento de texto (Chunking)
langchain-core
langchain-text-splitters